refactor: move map related functions
